+++ src/containerapp/azext_containerapp/_utils.py
@@ -1059,8 +1059,11 @@ def safe_get(model, *keys, default=None):
     if not model:
         return default
     for k in keys[:-1]:
-        model = model.get(k, {})
-    return model.get(keys[-1], default)
+        model = model.get(k)
+        if model is None:
+            return default
+    value = model.get(keys[-1], default)
+    return default if not value else value
 
 
 def safe_set(model, *keys, value):