+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_compose_basic.py
@@ -34,7 +34,7 @@ services:
             'compose': compose_file_name,
         })
 
-        create_containerapp_env(self, env_name, resource_group)
+        create_containerapp_env(self, env_name, resource_group, 'eastus')
 
         command_string = 'containerapp compose create'
         command_string += ' --compose-file-path {compose}'