+++ src/containerapp/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+0.3.18
+++++++
+* Fix "'NoneType' object has no attribute 'get'" error in `az containerapp up` with no ingress arguments
+
+0.3.17
+++++++
+* Fix polling logic for long running operations.
 
 0.3.16
 ++++++